faas-function/requirements.txt
minio
Pillow
moviepy